+++ crates/gqlitedb/src/value/contains.rs
@@ -1,21 +1,24 @@
 use crate::prelude::*;
 
-/// Compute contains, according to OpenCypher specification, specifically handling the comparison with null.
-pub(crate) fn contains(
-  container: &Vec<graph::Value>,
-  value: &graph::Value,
-) -> value::ComparisonResult
+pub(crate) enum ContainResult
 {
-  use value::ComparisonResult;
+  True,
+  False,
+  ComparedNull,
+}
+
+/// Compute contains, according to OpenCypher specification, specifically handling the comparison with null.
+pub(crate) fn contains(container: &Vec<graph::Value>, value: &graph::Value) -> ContainResult
+{
   if value.is_null()
   {
     if container.is_empty()
     {
-      ComparisonResult::False
+      ContainResult::False
     }
     else
     {
-      ComparisonResult::ComparedNull
+      ContainResult::ComparedNull
     }
   }
   else
@@ -23,22 +26,22 @@ pub(crate) fn contains(
     let mut has_compared_to_null = false;
     for v_c in container.iter()
     {
-      use value::ComparisonResult;
+      use value::ContainResult;
       match value::compare(v_c, &value)
       {
-        ComparisonResult::True => return ComparisonResult::True,
-        ComparisonResult::False =>
+        value::Ordering::Equal => return ContainResult::True,
+        value::Ordering::ComparedNull => has_compared_to_null = true,
+        _ =>
         {}
-        ComparisonResult::ComparedNull => has_compared_to_null = true,
       }
     }
     if has_compared_to_null
     {
-      ComparisonResult::ComparedNull
+      ContainResult::ComparedNull
     }
     else
     {
-      ComparisonResult::False
+      ContainResult::False
     }
   }
 }